+++ sources/lib/run-time/mps-collector.c
@@ -533,8 +533,8 @@ void MMFreeMisc(void *old, size_t size)
   void *mfill = (fill);  \
     \
   mps_bool_t mrq = (rq);   /* init repeated slot size */  \
-  int mrep_size = (rep_size);  \
-  int mrep_size_slot = (rep_size_slot);  \
+  size_t mrep_size = (rep_size);  \
+  size_t mrep_size_slot = (rep_size_slot);  \
     \
   mps_bool_t mrfq = (rfq);  /* init word repeated slot data */  \
   type mword_fill = (type)(word_fill);  \
@@ -635,8 +635,8 @@ void *primitive_alloc_s(size_t size,
 RUN_TIME_API
 void *primitive_alloc_r(size_t size,
                         void *wrapper,
-                        int rep_size,
-                        int rep_size_slot)
+                        size_t rep_size,
+                        size_t rep_size_slot)
 {
   alloc_internal(size, wrapper,
                  0, 0, 0, 0,
@@ -650,8 +650,8 @@ void *primitive_alloc_r(size_t size,
 RUN_TIME_API
 void *primitive_alloc_rf(size_t size,
                          void *wrapper,
-                         int rep_size,
-                         int rep_size_slot,
+                         size_t rep_size,
+                         size_t rep_size_slot,
                          dylan_object fill)
 {
   alloc_internal(size, wrapper,
@@ -669,8 +669,8 @@ void *primitive_alloc_s_r(size_t size,
                           void *wrapper,
                           int no_to_fill,
                           void *fill,
-                          int rep_size,
-                          int rep_size_slot)
+                          size_t rep_size,
+                          size_t rep_size_slot)
 {
   alloc_internal(size, wrapper,
                  0, 0, 0, 0,
@@ -688,8 +688,8 @@ void *primitive_alloc_s_ ## suffix(size_t size, \
                                    void *wrapper, \
                                    int no_to_fill, \
                                    void *fill, \
-                                   int rep_size, \
-                                   int rep_size_slot, \
+                                   size_t rep_size, \
+                                   size_t rep_size_slot, \
                                    type rep_fill) \
 { \
   alloc_internal(size, wrapper,  \
@@ -713,8 +713,8 @@ void *primitive_alloc_s_rbf(size_t size,
                             void *wrapper,
                             int no_to_fill,
                             void *fill,
-                            int rep_size,
-                            int rep_size_slot,
+                            size_t rep_size,
+                            size_t rep_size_slot,
                             int byte_fill)
 {
   alloc_internal(size, wrapper,
@@ -732,8 +732,8 @@ void *primitive_alloc_s_rbfz(size_t size,
                              void *wrapper,
                              int no_to_fill,
                              void *fill,
-                             int rep_size,
-                             int rep_size_slot,
+                             size_t rep_size,
+                             size_t rep_size_slot,
                              int byte_fill)
 {
   alloc_internal(size, wrapper,
@@ -749,8 +749,8 @@ void *primitive_alloc_s_rbfz(size_t size,
 RUN_TIME_API
 void *primitive_alloc_rbfz(size_t size,
                            void *wrapper,
-                           int rep_size,
-                           int rep_size_slot,
+                           size_t rep_size,
+                           size_t rep_size_slot,
                            int byte_fill)
 {
   alloc_internal(size, wrapper,
@@ -768,8 +768,8 @@ void *primitive_alloc_s_rb(size_t size,
                            void *wrapper,
                            int no_to_fill,
                            void *fill,
-                           int rep_size,
-                           int rep_size_slot)
+                           size_t rep_size,
+                           size_t rep_size_slot)
 {
   alloc_internal(size, wrapper,
                  0, 0, 0, 0,
@@ -800,8 +800,8 @@ void *primitive_alloc_leaf_s_r(size_t size,
                                void *wrapper,
                                int no_to_fill,
                                void *fill,
-                               int rep_size,
-                               int rep_size_slot)
+                               size_t rep_size,
+                               size_t rep_size_slot)
 {
   alloc_internal(size, wrapper,
                  0, 0, 0, 0,
@@ -864,8 +864,8 @@ void *primitive_alloc_leaf_s(size_t size,
 RUN_TIME_API
 void *primitive_alloc_leaf_r(size_t size,
                              void *wrapper,
-                             int rep_size,
-                             int rep_size_slot)
+                             size_t rep_size,
+                             size_t rep_size_slot)
 {
   alloc_internal(size, wrapper,
                  0, 0, 0, 0,
@@ -882,8 +882,8 @@ void *primitive_alloc_leaf_s_rbf(size_t size,
                                  void *wrapper,
                                  int no_to_fill,
                                  void *fill,
-                                 int rep_size,
-                                 int rep_size_slot,
+                                 size_t rep_size,
+                                 size_t rep_size_slot,
                                  int byte_fill)
 {
   alloc_internal(size, wrapper,
@@ -899,8 +899,8 @@ void *primitive_alloc_leaf_s_rbf(size_t size,
 RUN_TIME_API \
 void *primitive_alloc_leaf_ ## suffix(size_t size, \
                                       void *wrapper, \
-                                      int rep_size, \
-                                      int rep_size_slot, \
+                                      size_t rep_size, \
+                                      size_t rep_size_slot, \
                                       type rep_fill) \
 { \
   alloc_internal(size, wrapper,  \
@@ -925,8 +925,8 @@ void *primitive_alloc_leaf_s_rbfz(size_t size,
                                   void *wrapper,
                                   int no_to_fill,
                                   void *fill,
-                                  int rep_size,
-                                  int rep_size_slot,
+                                  size_t rep_size,
+                                  size_t rep_size_slot,
                                   int byte_fill)
 {
   alloc_internal(size, wrapper,
@@ -942,8 +942,8 @@ void *primitive_alloc_leaf_s_rbfz(size_t size,
 RUN_TIME_API
 void *primitive_alloc_leaf_rbfz(size_t size,
                                 void *wrapper,
-                                int rep_size,
-                                int rep_size_slot,
+                                size_t rep_size,
+                                size_t rep_size_slot,
                                 int byte_fill)
 {
   alloc_internal(size, wrapper,
@@ -961,8 +961,8 @@ void *primitive_alloc_leaf_s_rb(size_t size,
                                 void *wrapper,
                                 int no_to_fill,
                                 void *fill,
-                                int rep_size,
-                                int rep_size_slot)
+                                size_t rep_size,
+                                size_t rep_size_slot)
 {
   alloc_internal(size, wrapper,
                  0, 0, 0, 0,
@@ -980,8 +980,8 @@ void *primitive_alloc_exact_awl_s_r(size_t size,
                                     void *assoc,
                                     int no_to_fill,
                                     void *fill,
-                                    int rep_size,
-                                    int rep_size_slot)
+                                    size_t rep_size,
+                                    size_t rep_size_slot)
 {
   alloc_internal(size, wrapper,
                  1, assoc, 0, 0,
@@ -999,8 +999,8 @@ void *primitive_alloc_weak_awl_s_r(size_t size,
                                    void *assoc,
                                    int no_to_fill,
                                    void *fill,
-                                   int rep_size,
-                                   int rep_size_slot)
+                                   size_t rep_size,
+                                   size_t rep_size_slot)
 {
   alloc_internal(size, wrapper,
                  1, assoc, 0, 0,
@@ -1016,8 +1016,8 @@ RUN_TIME_API
 void *primitive_alloc_exact_awl_rf(size_t size,
                                    void *wrapper,
                                    void *assoc,
-                                   int rep_size,
-                                   int rep_size_slot,
+                                   size_t rep_size,
+                                   size_t rep_size_slot,
                                    void *fill)
 {
   alloc_internal(size, wrapper,
@@ -1034,8 +1034,8 @@ RUN_TIME_API
 void *primitive_alloc_weak_awl_rf(size_t size,
                                   void *wrapper,
                                   void *assoc,
-                                  int rep_size,
-                                  int rep_size_slot,
+                                  size_t rep_size,
+                                  size_t rep_size_slot,
                                   void *fill)
 {
   alloc_internal(size, wrapper,
@@ -1053,8 +1053,8 @@ void *primitive_alloc_wrapper_s_r(size_t size,
                                   void *wrapper,
                                   int no_to_fill,
                                   void *fill,
-                                  int rep_size,
-                                  int rep_size_slot)
+                                  size_t rep_size,
+                                  size_t rep_size_slot)
 {
   alloc_internal(size, wrapper,
                  0, 0, 0, 0,
@@ -1069,8 +1069,8 @@ void *primitive_alloc_wrapper_s_r(size_t size,
 RUN_TIME_API
 void *primitive_alloc_rt(size_t size,
                          void *wrapper,
-                         int rep_size,
-                         int rep_size_slot,
+                         size_t rep_size,
+                         size_t rep_size_slot,
                          void *template)
 {
   void **object;
@@ -1115,8 +1115,8 @@ void *primitive_copy(size_t size,
 
 RUN_TIME_API
 void *primitive_copy_r(size_t size,
-                       int rep_size,
-                       int rep_size_slot,
+                       size_t rep_size,
+                       size_t rep_size_slot,
                        void *template)
 {
   void **object;
